refactor @Controller initializer
