+++ src/decorators/controller/controller.ts
@@ -1,18 +1,10 @@
-import normalizePath from "normalize-path";
 import { container as defaultContainer } from "../../container/container";
 import { keys } from "../../container/keys";
-import { getGlobal } from "../../container/utils/getGlobal";
 import { validateContainer } from "../../container/utils/validateContainer";
-import { AnnotatedCacheStorage } from "../../interfaces/annotatedCacheStorage";
-import type { AnnotatedRouter } from "../../interfaces/annotatedRouter";
-import { Cache } from "../../interfaces/cache";
-import { setInjectables } from "../inject/setInjectables";
-import type { Class, ClassDecorator, HttpMethodMetadata } from "../types";
+import type { Class, ClassDecorator } from "../types";
 import { addClassToContainer } from "../utils/addClassToContainer";
-import { getMetadata } from "../utils/getMetadata";
-import { getMetadataProperty } from "../utils/getMetadataProperty";
 import { validateKind } from "../utils/validateKind";
-import { MetadataProperties } from "./metadataProperties";
+import { getInitializer } from "./utils/getInitializer";
 
 /**
  * A class decorator that encapsulates API endpoints
@@ -87,66 +79,8 @@ export const Controller = <T extends Class<object>>(
       throw new Error("Controller path argument is an empty string");
     }
 
-    context.addInitializer(function () {
-      const controller = new this();
-
-      const metadata = getMetadata(annotationName, context);
-      setInjectables(container, controller, metadata);
-
-      const methods = <Array<HttpMethodMetadata>>(
-        getMetadataProperty(metadata, MetadataProperties.methods, [])
-      );
-
-      const controllerMethodMetadata = methods.map((methodMetadata) => ({
-        ...methodMetadata,
-        handler: methodMetadata.handler.bind(controller),
-      }));
-
-      let router: AnnotatedRouter = getGlobal(container, keys.router);
-      const cacheStorage: AnnotatedCacheStorage = getGlobal(
-        container,
-        keys.cacheStorage,
-      );
-      for (const {
-        path: methodPath,
-        handler,
-        httpMethod,
-        cacheName,
-      } of controllerMethodMetadata) {
-        const routerMethod = <Exclude<keyof AnnotatedRouter, "handle">>(
-          httpMethod.toLowerCase()
-        );
-
-        if (typeof router[routerMethod] !== "function") {
-          throw new Error(
-            `Router is improperly configured. It should include ${routerMethod} method`,
-          );
-        }
-
-        router = router[routerMethod](
-          normalizePath("/" + [controllerPath, methodPath].join("/"), true),
-          async (request: Request) => {
-            let cache: Cache | undefined;
-            if (cacheName) {
-              cache = await cacheStorage.open(cacheName);
-
-              const response = await cache.match(request);
-              if (response) {
-                return response;
-              }
-            }
-
-            const response = await handler.call(controller, request);
-
-            if (cache) {
-              await cache.put(request, response);
-            }
-
-            return response;
-          },
-        );
-      }
-    });
-
+    context.addInitializer(
+      getInitializer(container, annotationName, context, controllerPath),
+    );
     addClassToContainer(container, keys.controllerClasses, constructor);
   }) as ClassDecorator<T>;
